bupdate

+++ src/base16.h
@@ -21,6 +21,7 @@
 #ifndef Base16_H
 #define Base16_H
 
+#include <cstdint>
 #include <algorithm>
 #include <string>
 #include <sstream>
@@ -45,7 +46,7 @@ public:
     /// \brief Map for fast lookup corresponding character
     /// \see Base64::kDecodeMap
     ///
-    static const std::unordered_map<int, int> kDecodeMap;
+    static const std::unordered_map<uint8_t, uint8_t> kDecodeMap;
 
     ///
     /// \brief Encodes input of length to hex encoding
